"完善record接口、解决mysql自增异常"

+++ src/main/java/cn/krl/visiteducationbackend/dto/AdminDTO.java
@@ -15,7 +15,7 @@ import javax.validation.constraints.NotNull;
 @Api("管理员传输对象")
 public class AdminDTO {
     @NotNull
-    @ApiModelProperty(name = "id",value = "管理员id",dataType = "Long")
+    @ApiModelProperty(name = "id",value = "管理员id",dataType = "Integer")
     private Integer id;
 
     @NotNull
